+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Temporarily don't fail when Chart.yaml doesn't have the config annotation in `architect create argoapp`.
+
 ## [3.5.2] - 2021-05-17
 
 ### Fixed